+++ tencentcloud/cdb/v20170320/models.go
@@ -1067,7 +1067,7 @@ type CreateDBImportJobRequest struct {
 	// 实例的 ID，格式如：cdb-c1nl9rpv，与云数据库控制台页面中显示的实例 ID 相同。
 	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`
 
-	// 文件名称。该文件是指用户已上传到腾讯云的文件。
+	// 文件名称。该文件是指用户已上传到腾讯云的文件，仅支持.sql文件。
 	FileName *string `json:"FileName,omitempty" name:"FileName"`
 
 	// 云数据库的用户名。