+++ tencentcloud/scf/v20180416/errors.go
@@ -26,6 +26,9 @@ const (
 	// 创建触发器失败。
 	FAILEDOPERATION_APIGW = "FailedOperation.Apigw"
 
+	// 获取Apm InstanceId失败。
+	FAILEDOPERATION_APMCONFIGINSTANCEID = "FailedOperation.ApmConfigInstanceId"
+
 	// 当前异步事件状态不支持此操作，请稍后重试。
 	FAILEDOPERATION_ASYNCEVENTSTATUS = "FailedOperation.AsyncEventStatus"
 
@@ -193,6 +196,15 @@ const (
 
 	// ApiGateway参数错误。
 	INVALIDPARAMETERVALUE_APIGATEWAY = "InvalidParameterValue.ApiGateway"
+
+	// ApmConfig参数传入错误。
+	INVALIDPARAMETERVALUE_APMCONFIG = "InvalidParameterValue.ApmConfig"
+
+	// ApmConfigInstanceId参数传入错误。
+	INVALIDPARAMETERVALUE_APMCONFIGINSTANCEID = "InvalidParameterValue.ApmConfigInstanceId"
+
+	// ApmConfigRegion参数传入错误。
+	INVALIDPARAMETERVALUE_APMCONFIGREGION = "InvalidParameterValue.ApmConfigRegion"
 
 	// 函数异步重试配置参数无效。
 	INVALIDPARAMETERVALUE_ASYNCTRIGGERCONFIG = "InvalidParameterValue.AsyncTriggerConfig"
@@ -595,6 +607,9 @@ const (
 
 	// 获取cfs信息错误。
 	RESOURCENOTFOUND_GETCFSNOTMATCH = "ResourceNotFound.GetCfsNotMatch"
+
+	// 未找到指定的ImageConfig，请创建后再试。
+	RESOURCENOTFOUND_IMAGECONFIG = "ResourceNotFound.ImageConfig"
 
 	// layer不存在。
 	RESOURCENOTFOUND_LAYER = "ResourceNotFound.Layer"